NFZ_Server/requirements.txt
cryptography==42.0.5
Flask==2.3.2
flask-cors==4.0.0
google-api-python-client==2.117.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
requests==2.31.0
scikit-learn==1.5.0
